chore: clean up PR #932 (Close #929 Trace app startup time)
