+++ src/dotnet/App.Server/Program.cs
@@ -85,13 +85,4 @@ internal static class Program
         static void TraceOutput(string m)
             => Console.WriteLine(m);
     }
-
-    internal class RootTraceAccessor : ITraceAccessor
-    {
-        public static readonly RootTraceAccessor Instance = new RootTraceAccessor();
-
-        private RootTraceAccessor() { }
-
-        public ITraceSession? Trace { get; set; }
-    }
 }
