+++ src/dotnet/App.Server/RootTraceAccessor.cs
@@ -1,0 +1,10 @@
+namespace ActualChat.App.Server;
+
+internal class RootTraceAccessor : ITraceAccessor
+{
+    public static readonly RootTraceAccessor Instance = new RootTraceAccessor();
+
+    private RootTraceAccessor() { }
+
+    public ITraceSession? Trace { get; set; }
+}
